Remove facility because it's deprecated
The GELF specification[1] allows a top-level `facility` field to be
specified, but this is deprecated and it should be sent as `_facility`
(an "additional" field) instead if anybody still wants to use it. The
original purpose of the field seems to have been a sort of hangover
from the RFC 5452 syslog facilities (integers 0-23) that never was
going in the right direction.

Thus, we remove any reference to facilities; clients that want to log
it should add it as an additional field as the GELF spec says.

[1]: http://docs.graylog.org/en/2.2/pages/gelf.html#gelf-payload-specification